requirements.txt
feedparser
lxml
marshmallow
nameko
nameko-sqlalchemy
psycopg2-binary
PyJWT
requests
SQLAlchemy
Werkzeug
